refactor: Create auth package and move AuthPayload model and creating token to it
