+++ auth/auth.go
@@ -1,4 +1,4 @@
-package models
+package auth
 
 type AuthPayload struct {
 	Token string `json:"token"`
